DevasterV4.py
This Python script executes a DNS hijacking attack. It prompts the user to input the target IP address or domain, the interval between queries, and the spoofed sender IP address. Then, it continuously sends crafted DNS query packets to the target, simulating a flood of requests, potentially disrupting the target's DNS resolution. Finally, it gracefully handles interruptions such as keyboard interrupts, allowing the user to terminate the attack.